+++ zakat/dashboard/templates/auth/login.html
@@ -19,7 +19,7 @@
 {% endif %}
 
 
-    <form action="." method="post">
+    <form action="../../../accounts/templates/accounts" method="post">
         {{ form.as_p }}
         {% csrf_token %}
         <p><input type="submit" value="Log-in"></p>
